more modularization and refactoring.
